refactor: move `validate_field` to it's own module
