+++ src/blueprint/mod.rs
@@ -5,6 +5,7 @@ mod from_config;
 mod into_schema;
 mod operation;
 mod timeout;
+mod validate;
 
 pub use blueprint::*;
 pub use const_utils::*;
